Added images and style etc
Have basic idea set up including Galactic Cruise Lines ad.  However, I'm stuck on how to get the layout to apply and be recognized by the browser, for pages other than the first one.  This is critical.  Instructor video shows using older pages as if they were part of the layout.  The error I get is that there is no Doctype HTML etc for the new pages that have the layout header...